Revert "Introduce privateApiElements variant"
