+++ platforms/documentation/docs/src/docs/userguide/dep-man/04-modeling-features/variant_model.adoc
@@ -355,9 +355,120 @@ It also accepts the `--all` flag to include information about legacy and depreca
 
 Here is the output of the `outgoingVariants` task on a freshly generated `java-library` project:
 
-.Outgoing variants report
+[listing]
 ----
-include::{snippetsPath}/java-library/quickstart/tests/outgoingVariants.out[]
+> Task :outgoingVariants
+--------------------------------------------------
+Variant apiElements
+--------------------------------------------------
+API elements for the 'main' feature.
+
+Capabilities
+    - new-java-library:lib:unspecified (default capability)
+Attributes
+    - org.gradle.category            = library
+    - org.gradle.dependency.bundling = external
+    - org.gradle.jvm.version         = 11
+    - org.gradle.libraryelements     = jar
+    - org.gradle.usage               = java-api
+Artifacts
+    - build/libs/lib.jar (artifactType = jar)
+
+Secondary Variants (*)
+
+    --------------------------------------------------
+    Secondary Variant classes
+    --------------------------------------------------
+        Description = Directories containing compiled class files for main.
+
+        Attributes
+            - org.gradle.category            = library
+            - org.gradle.dependency.bundling = external
+            - org.gradle.jvm.version         = 11
+            - org.gradle.libraryelements     = classes
+            - org.gradle.usage               = java-api
+        Artifacts
+            - build/classes/java/main (artifactType = java-classes-directory)
+
+--------------------------------------------------
+Variant mainSourceElements (i)
+--------------------------------------------------
+Description = List of source directories contained in the Main SourceSet.
+
+Capabilities
+    - new-java-library:lib:unspecified (default capability)
+Attributes
+    - org.gradle.category            = verification
+    - org.gradle.dependency.bundling = external
+    - org.gradle.verificationtype    = main-sources
+Artifacts
+    - src/main/java (artifactType = directory)
+    - src/main/resources (artifactType = directory)
+
+--------------------------------------------------
+Variant runtimeElements
+--------------------------------------------------
+Runtime elements for the 'main' feature.
+
+Capabilities
+    - new-java-library:lib:unspecified (default capability)
+Attributes
+    - org.gradle.category            = library
+    - org.gradle.dependency.bundling = external
+    - org.gradle.jvm.version         = 11
+    - org.gradle.libraryelements     = jar
+    - org.gradle.usage               = java-runtime
+Artifacts
+    - build/libs/lib.jar (artifactType = jar)
+
+Secondary Variants (*)
+
+    --------------------------------------------------
+    Secondary Variant classes
+    --------------------------------------------------
+        Description = Directories containing compiled class files for main.
+
+        Attributes
+            - org.gradle.category            = library
+            - org.gradle.dependency.bundling = external
+            - org.gradle.jvm.version         = 11
+            - org.gradle.libraryelements     = classes
+            - org.gradle.usage               = java-runtime
+        Artifacts
+            - build/classes/java/main (artifactType = java-classes-directory)
+
+    --------------------------------------------------
+    Secondary Variant resources
+    --------------------------------------------------
+        Description = Directories containing the project's assembled resource files for use at runtime.
+
+        Attributes
+            - org.gradle.category            = library
+            - org.gradle.dependency.bundling = external
+            - org.gradle.jvm.version         = 11
+            - org.gradle.libraryelements     = resources
+            - org.gradle.usage               = java-runtime
+        Artifacts
+            - build/resources/main (artifactType = java-resources-directory)
+
+--------------------------------------------------
+Variant testResultsElementsForTest (i)
+--------------------------------------------------
+Description = Directory containing binary results of running tests for the test Test Suite's test target.
+
+Capabilities
+    - new-java-library:lib:unspecified (default capability)
+Attributes
+    - org.gradle.category              = verification
+    - org.gradle.testsuite.name        = test
+    - org.gradle.testsuite.target.name = test
+    - org.gradle.testsuite.type        = unit-test
+    - org.gradle.verificationtype      = test-results
+Artifacts
+    - build/test-results/test/binary (artifactType = directory)
+
+(i) Configuration uses incubating attributes such as Category.VERIFICATION.
+(*) Secondary variants are variants created via the Configuration#getOutgoing(): ConfigurationPublications API which also participate in selection, in addition to the configuration itself.
 ----
 
 From this you can see the two main variants that are exposed by a java library, `apiElements` and `runtimeElements`.
