+++ platforms/documentation/docs/src/docs/userguide/jvm/java_library_plugin.adoc
@@ -193,16 +193,10 @@ The role of each configuration is described in the following tables:
 | Configuration name | Role | Consumable? | Resolvable? | Description
 
 | `apiElements`
-| For compiling against the public API of this library
+| For compiling against this library
 | yes
 | no
 | This configuration is meant to be used by consumers, to retrieve all the elements necessary to compile against this library.
-
-| `privateApiElements`
-| For compiling against the private API of this library
-| yes
-| no
-| This configuration is meant to be used by consumers, to support unit test dependencies during compilation.
 
 | `runtimeElements`
 | For executing this library
