+++ platforms/documentation/docs/src/snippets/dependencyManagement/inspectingDependencies-dependencyReason/tests/dependencyReasonReport.out
@@ -6,7 +6,6 @@ org.ow2.asm:asm:7.1
     | org.gradle.category            | library  | library      |
     | org.gradle.libraryelements     | jar      | classes      |
     | org.gradle.usage               | java-api | java-api     |
-    | org.gradle.api-type            |          | public       |
     | org.gradle.dependency.bundling |          | external     |
     | org.gradle.jvm.environment     |          | standard-jvm |
     | org.gradle.jvm.version         |          | 11           |
