+++ platforms/documentation/docs/src/snippets/dependencyManagement/modelingFeatures-crossProjectPublications-advanced/tests/outgoingVariants.out
@@ -2,7 +2,7 @@
 --------------------------------------------------
 Variant apiElements
 --------------------------------------------------
-Public API elements for the 'main' feature.
+API elements for the 'main' feature.
 
 Attributes
     - org.gradle.category            = library
