+++ platforms/documentation/docs/src/snippets/java-library/quickstart/kotlin/build.gradle.kts
@@ -17,11 +17,3 @@ dependencies {
     implementation("org.apache.commons:commons-lang3:3.5")
 }
 // end::dependencies[]
-
-if (hasProperty("toolchainJvmVersion")) {
-    java {
-        toolchain {
-            languageVersion = JavaLanguageVersion.of(property("toolchainJvmVersion") as String)
-        }
-    }
-}
