--- platforms/documentation/docs/src/snippets/java-library/quickstart/tests/javaLibraryOutgoingVariants.sample.conf
@@ -1,4 +1,0 @@
-executable: gradle
-args: outgoingVariants
-flags: --quiet "-PtoolchainJvmVersion=11" # Force java 11 to get consistent output
-expected-output-file: outgoingVariants.out
