--- platforms/documentation/docs/src/snippets/java-library/quickstart/tests/outgoingVariants.out
@@ -1,168 +1,0 @@
---------------------------------------------------
-Variant apiElements (i)
---------------------------------------------------
-Public API elements for the 'main' feature.
-
-Capabilities
-    - :quickstart:unspecified (default capability)
-Attributes
-    - org.gradle.api-type            = public
-    - org.gradle.category            = library
-    - org.gradle.dependency.bundling = external
-    - org.gradle.jvm.version         = 11
-    - org.gradle.libraryelements     = jar
-    - org.gradle.usage               = java-api
-Artifacts
-    - build/libs/quickstart.jar (artifactType = jar)
-
-Secondary Variants (*)
-
-    --------------------------------------------------
-    Secondary Variant classes (i)
-    --------------------------------------------------
-    Directories containing compiled class files for main.
-
-    Attributes
-        - org.gradle.api-type            = public
-        - org.gradle.category            = library
-        - org.gradle.dependency.bundling = external
-        - org.gradle.jvm.version         = 11
-        - org.gradle.libraryelements     = classes
-        - org.gradle.usage               = java-api
-    Artifacts
-        - build/classes/java/main (artifactType = java-classes-directory)
-
---------------------------------------------------
-Variant archives
---------------------------------------------------
-Configuration for archive artifacts.
-
-Capabilities
-    - :quickstart:unspecified (default capability)
-Artifacts
-    - build/libs/quickstart.jar (artifactType = jar)
-
---------------------------------------------------
-Variant default
---------------------------------------------------
-Configuration for default artifacts.
-
-Capabilities
-    - :quickstart:unspecified (default capability)
-Artifacts
-    - build/libs/quickstart.jar (artifactType = jar)
-
---------------------------------------------------
-Variant mainSourceElements (i)
---------------------------------------------------
-List of source directories contained in the Main SourceSet.
-
-Capabilities
-    - :quickstart:unspecified (default capability)
-Attributes
-    - org.gradle.category            = verification
-    - org.gradle.dependency.bundling = external
-    - org.gradle.verificationtype    = main-sources
-Artifacts
-    - src/main/java (artifactType = directory)
-    - src/main/resources (artifactType = directory)
-
---------------------------------------------------
-Variant privateApiElements (i)
---------------------------------------------------
-Private API elements for the 'main' feature.
-
-Capabilities
-    - :quickstart:unspecified (default capability)
-Attributes
-    - org.gradle.api-type            = private
-    - org.gradle.category            = library
-    - org.gradle.dependency.bundling = external
-    - org.gradle.jvm.version         = 11
-    - org.gradle.libraryelements     = jar
-    - org.gradle.usage               = java-api
-Artifacts
-    - build/libs/quickstart.jar (artifactType = jar)
-
-Secondary Variants (*)
-
-    --------------------------------------------------
-    Secondary Variant classesAndResources (i)
-    --------------------------------------------------
-    Directories containing compiled class files and resources for main.
-
-    Attributes
-        - org.gradle.api-type            = private
-        - org.gradle.category            = library
-        - org.gradle.dependency.bundling = external
-        - org.gradle.jvm.version         = 11
-        - org.gradle.libraryelements     = classes+resources
-        - org.gradle.usage               = java-api
-    Artifacts
-        - build/classes/java/main (artifactType = directory)
-        - build/resources/main (artifactType = directory)
-
---------------------------------------------------
-Variant runtimeElements
---------------------------------------------------
-Runtime elements for the 'main' feature.
-
-Capabilities
-    - :quickstart:unspecified (default capability)
-Attributes
-    - org.gradle.category            = library
-    - org.gradle.dependency.bundling = external
-    - org.gradle.jvm.version         = 11
-    - org.gradle.libraryelements     = jar
-    - org.gradle.usage               = java-runtime
-Artifacts
-    - build/libs/quickstart.jar (artifactType = jar)
-
-Secondary Variants (*)
-
-    --------------------------------------------------
-    Secondary Variant classes
-    --------------------------------------------------
-    Directories containing compiled class files for main.
-
-    Attributes
-        - org.gradle.category            = library
-        - org.gradle.dependency.bundling = external
-        - org.gradle.jvm.version         = 11
-        - org.gradle.libraryelements     = classes
-        - org.gradle.usage               = java-runtime
-    Artifacts
-        - build/classes/java/main (artifactType = java-classes-directory)
-
-    --------------------------------------------------
-    Secondary Variant resources
-    --------------------------------------------------
-    Directories containing assembled resource files for main.
-
-    Attributes
-        - org.gradle.category            = library
-        - org.gradle.dependency.bundling = external
-        - org.gradle.jvm.version         = 11
-        - org.gradle.libraryelements     = resources
-        - org.gradle.usage               = java-runtime
-    Artifacts
-        - build/resources/main (artifactType = java-resources-directory)
-
---------------------------------------------------
-Variant testResultsElementsForTest (i)
---------------------------------------------------
-Directory containing binary results of running tests for the test Test Suite's test target.
-
-Capabilities
-    - :quickstart:unspecified (default capability)
-Attributes
-    - org.gradle.category              = verification
-    - org.gradle.testsuite.name        = test
-    - org.gradle.testsuite.target.name = test
-    - org.gradle.testsuite.type        = unit-test
-    - org.gradle.verificationtype      = test-results
-Artifacts
-    - build/test-results/test/binary (artifactType = directory)
-
-(i) Configuration uses incubating attributes such as Category.VERIFICATION.
-(*) Secondary variants are variants created via the Configuration#getOutgoing(): ConfigurationPublications API which also participate in selection, in addition to the configuration itself.
