+++ platforms/jvm/java-platform/build.gradle.kts
@@ -52,10 +52,6 @@ dependencies {
         because("need to access JavaPluginExtension")
     }
 
-    testRuntimeOnly(project(":distributions-core")) {
-        because("ProjectBuilder tests load services from a Gradle distribution.")
-    }
-
     testImplementation(testFixtures(project(":core")))
 
     integTestDistributionRuntimeOnly(project(":distributions-jvm"))
