+++ platforms/jvm/java-platform/src/integTest/groovy/org/gradle/integtests/resolve/platforms/NativeAlignmentWithJavaPlatformResolveIntegrationTest.groovy
@@ -204,7 +204,7 @@ class NativeAlignmentWithJavaPlatformResolveIntegrationTest extends AbstractModu
                         'org.gradle.status':'release',
                         'org.gradle.usage': 'java-api',
                         'org.gradle.libraryelements': 'jar',
-                        'org.gradle.api-type': 'public']
+                    ]
                     module('com.acme.foo:platform:1.1') {
                         variant "apiElements", [
                             'org.gradle.category':'platform',
@@ -223,7 +223,6 @@ class NativeAlignmentWithJavaPlatformResolveIntegrationTest extends AbstractModu
                             'org.gradle.status':'release',
                             'org.gradle.usage': 'java-api',
                             'org.gradle.libraryelements': 'jar',
-                            'org.gradle.api-type': 'public'
                         ]
                         byConflictResolution("between versions 1.1 and 1.0")
                         byConstraint("platform alignment")
@@ -237,7 +236,7 @@ class NativeAlignmentWithJavaPlatformResolveIntegrationTest extends AbstractModu
                         'org.gradle.status':'release',
                         'org.gradle.usage': 'java-api',
                         'org.gradle.libraryelements': 'jar',
-                        'org.gradle.api-type': 'public']
+                    ]
                     module('com.acme.foo:platform:1.1') {
                         variant "apiElements", [
                             'org.gradle.category':'platform',
