+++ platforms/jvm/java-platform/src/test/java/org/gradle/api/plugins/JavaPlatformPluginTest.groovy
