+++ platforms/jvm/platform-jvm/src/main/java/org/gradle/api/plugins/jvm/internal/DefaultJvmPluginServices.java
@@ -63,7 +63,7 @@ public class DefaultJvmPluginServices implements JvmPluginServices {
     public void configureAsCompileClasspath(HasConfigurableAttributes<?> configuration) {
         configureAttributes(
             configuration,
-            details -> details.library().apiUsage().withExternalDependencies().preferStandardJVM().publicApiType()
+            details -> details.library().apiUsage().withExternalDependencies().preferStandardJVM()
         );
     }
 
@@ -87,15 +87,7 @@ public class DefaultJvmPluginServices implements JvmPluginServices {
     public void configureAsApiElements(HasConfigurableAttributes<?> configuration) {
         configureAttributes(
             configuration,
-            details -> details.library().apiUsage().asJar().withExternalDependencies().publicApiType()
-        );
-    }
-
-    @Override
-    public void configureAsPrivateApiElements(HasConfigurableAttributes<?> configuration) {
-        configureAttributes(
-            configuration,
-            details -> details.library().apiUsage().asJar().withExternalDependencies().privateApiType()
+            details -> details.library().apiUsage().asJar().withExternalDependencies()
         );
     }
 
@@ -156,23 +148,6 @@ public class DefaultJvmPluginServices implements JvmPluginServices {
                 .map(file -> new LazyJavaDirectoryArtifact(project.getTaskDependencyFactory(), ArtifactTypeDefinition.JVM_CLASS_DIRECTORY, classesDirs, providerFactory.provider(() -> file)))
                 .collect(Collectors.toList());
         });
-
-        return variant;
-    }
-
-    @Override
-    public ConfigurationVariant configureClassesAndResourcesDirectoryVariant(Configuration configuration, SourceSet sourceSet) {
-        ConfigurationPublications publications = configuration.getOutgoing();
-        ConfigurationVariantInternal variant = (ConfigurationVariantInternal) publications.getVariants().maybeCreate("classesAndResources");
-        variant.setDescription("Directories containing compiled class files and resources for " + sourceSet.getName() + ".");
-        variant.getAttributes().attribute(LibraryElements.LIBRARY_ELEMENTS_ATTRIBUTE, objectFactory.named(LibraryElements.class, LibraryElements.CLASSES_AND_RESOURCES));
-        variant.artifactsProvider(() -> {
-            FileCollection dirs = sourceSet.getOutput();
-            return dirs.getFiles().stream()
-                .map(file -> new LazyJavaDirectoryArtifact(project.getTaskDependencyFactory(), ArtifactTypeDefinition.DIRECTORY_TYPE, dirs, providerFactory.provider(() -> file)))
-                .collect(Collectors.toList());
-        });
-
         return variant;
     }
