+++ platforms/jvm/platform-jvm/src/main/java/org/gradle/api/plugins/jvm/internal/JvmPluginServices.java
@@ -40,20 +40,12 @@ public interface JvmPluginServices extends JvmEcosystemUtilities {
     ConfigurationVariant configureResourcesDirectoryVariant(Configuration configuration, SourceSet sourceSet);
 
     /**
-     * Registers a variant on {@code configuration} which exposes the classes defined by {@code sourceSet}.
+     * Registers a variant on {@code configuration} which exposes the classses defined by {@code sourceSet}.
      *
      * @param configuration The {@link Configuration} for which a classes variant should be exposed.
      * @param sourceSet The {@link SourceSet} which will contribute classes to this variant.
      */
     ConfigurationVariant configureClassesDirectoryVariant(Configuration configuration, SourceSet sourceSet);
-
-    /**
-     * Registers a variant on {@code configuration} which exposes the classes and resources defined by {@code sourceSet}.
-     *
-     * @param configuration The {@link Configuration} for which a classes and resources variant should be exposed.
-     * @param sourceSet The {@link SourceSet} which will contribute classes and resources to this variant.
-     */
-    ConfigurationVariant configureClassesAndResourcesDirectoryVariant(Configuration configuration, SourceSet sourceSet);
 
     /**
      * Configures a configuration with reasonable defaults to be resolved as a compile classpath.
@@ -63,18 +55,11 @@ public interface JvmPluginServices extends JvmEcosystemUtilities {
     void configureAsCompileClasspath(HasConfigurableAttributes<?> configuration);
 
     /**
-     * Configures a consumable configuration to provide a public API classpath.
+     * Configures a consumable configuration to provide an API compile classpath.
      *
      * @param configuration the configuration to be configured
      */
     void configureAsApiElements(HasConfigurableAttributes<?> configuration);
-
-    /**
-     * Configures a consumable configuration to provide a private API classpath.
-     *
-     * @param configuration the configuration to be configured
-     */
-    void configureAsPrivateApiElements(HasConfigurableAttributes<?> configuration);
 
     /**
      * Configures a consumable configuration to provide a runtime classpath.
