+++ platforms/jvm/plugins-groovy/src/integTest/groovy/org/gradle/groovy/GroovyJavaLibraryInteractionIntegrationTest.groovy
@@ -84,7 +84,7 @@ class GroovyJavaLibraryInteractionIntegrationTest extends AbstractDependencyReso
                         'org.gradle.jvm.version': JavaVersion.current().majorVersion,
                         'org.gradle.usage': 'java-api',
                         'org.gradle.libraryelements': 'jar',
-                        'org.gradle.api-type': 'public'])
+                        ])
                     switch (expected) {
                         case "jar":
                             artifact(name: "groovyLib")
