+++ platforms/jvm/plugins-java-base/src/integTest/groovy/org/gradle/api/plugins/jvm/internal/DefaultJvmFeatureTest.groovy
@@ -46,7 +46,6 @@ class DefaultJvmFeatureTest extends AbstractProjectBuilderSpec {
         feature.compileClasspathConfiguration == project.configurations.getByName(JvmConstants.COMPILE_CLASSPATH_CONFIGURATION_NAME)
         feature.runtimeElementsConfiguration == project.configurations.getByName(JvmConstants.RUNTIME_ELEMENTS_CONFIGURATION_NAME)
         feature.apiElementsConfiguration == project.configurations.getByName(JvmConstants.API_ELEMENTS_CONFIGURATION_NAME)
-        feature.privateApiElementsConfiguration == project.configurations.getByName(JvmConstants.PRIVATE_API_ELEMENTS_CONFIGURATION_NAME)
         feature.implementationConfiguration == project.configurations.getByName(JvmConstants.IMPLEMENTATION_CONFIGURATION_NAME)
         feature.runtimeOnlyConfiguration == project.configurations.getByName(JvmConstants.RUNTIME_ONLY_CONFIGURATION_NAME)
         feature.compileOnlyConfiguration == project.configurations.getByName(JvmConstants.COMPILE_ONLY_CONFIGURATION_NAME)
@@ -72,7 +71,6 @@ class DefaultJvmFeatureTest extends AbstractProjectBuilderSpec {
         feature.compileClasspathConfiguration == project.configurations.getByName('featureCompileClasspath')
         feature.runtimeElementsConfiguration == project.configurations.getByName('featureRuntimeElements')
         feature.apiElementsConfiguration == project.configurations.getByName('featureApiElements')
-        feature.privateApiElementsConfiguration == project.configurations.getByName('featurePrivateApiElements')
         feature.implementationConfiguration == project.configurations.getByName('featureImplementation')
         feature.runtimeOnlyConfiguration == project.configurations.getByName('featureRuntimeOnly')
         feature.compileOnlyConfiguration == project.configurations.getByName('featureCompileOnly')
