+++ platforms/jvm/plugins-java-base/src/main/java/org/gradle/api/plugins/jvm/internal/DefaultJvmFeature.java
@@ -105,7 +105,6 @@ public class DefaultJvmFeature implements JvmFeatureInternal {
     // Outgoing variants
     private final Configuration apiElements;
     private final Configuration runtimeElements;
-    private final Configuration privateApiElements;
 
     // Configurable outgoing variants
     private Configuration javadocElements;
@@ -157,7 +156,6 @@ public class DefaultJvmFeature implements JvmFeatureInternal {
         PublishArtifact jarArtifact = new LazyPublishArtifact(jar, project.getFileResolver(), project.getTaskDependencyFactory());
         this.apiElements = createApiElements(configurations, jarArtifact, compileJava, useMigrationRoleForElementsConfigurations);
         this.runtimeElements = createRuntimeElements(configurations, jarArtifact, compileJava, useMigrationRoleForElementsConfigurations);
-        this.privateApiElements = createPrivateApiElements(configurations, jarArtifact);
 
         if (extendProductionCode) {
             doExtendProductionCode();
@@ -233,7 +231,7 @@ public class DefaultJvmFeature implements JvmFeatureInternal {
         jvmLanguageUtilities.useDefaultTargetPlatformInference(apiElements, compileJava);
         jvmPluginServices.configureAsApiElements(apiElements);
         capabilities.forEach(apiElements.getOutgoing()::capability);
-        apiElements.setDescription("Public API elements for the '" + name + "' feature.");
+        apiElements.setDescription("API elements for the '" + name + "' feature.");
 
         // Configure variants
         addJarArtifactToConfiguration(apiElements, jarArtifact);
@@ -264,25 +262,6 @@ public class DefaultJvmFeature implements JvmFeatureInternal {
         jvmPluginServices.configureResourcesDirectoryVariant(runtimeElements, sourceSet);
 
         return runtimeElements;
-    }
-
-    private Configuration createPrivateApiElements(
-        RoleBasedConfigurationContainerInternal configurations,
-        PublishArtifact jarArtifact
-    ) {
-         String configName = getConfigurationName(JvmConstants.PRIVATE_API_ELEMENTS_CONFIGURATION_NAME);
-         return configurations.consumableUnlocked(configName, conf -> {
-            jvmLanguageUtilities.useDefaultTargetPlatformInference(conf, compileJava);
-            jvmPluginServices.configureAsPrivateApiElements(conf);
-            capabilities.forEach(conf.getOutgoing()::capability);
-            conf.setDescription("Private API elements for the '" + name + "' feature.");
-
-            conf.extendsFrom(implementation, compileOnly);
-
-            // Configure variants
-            addJarArtifactToConfiguration(conf, jarArtifact);
-            jvmPluginServices.configureClassesAndResourcesDirectoryVariant(conf, sourceSet);
-        });
     }
 
     private static Configuration maybeCreateElementsConfiguration(
@@ -460,11 +439,6 @@ public class DefaultJvmFeature implements JvmFeatureInternal {
     @Override
     public Configuration getApiElementsConfiguration() {
         return apiElements;
-    }
-
-    @Override
-    public Configuration getPrivateApiElementsConfiguration() {
-        return privateApiElements;
     }
 
     @Override
