+++ platforms/jvm/plugins-java-library/src/integTest/groovy/org/gradle/java/JavaLibraryCrossProjectTargetJvmVersionIntegrationTest.groovy
@@ -18,6 +18,7 @@ package org.gradle.java
 
 import org.gradle.integtests.fixtures.AbstractIntegrationSpec
 import org.gradle.integtests.fixtures.resolve.ResolveTestFixture
+import org.hamcrest.Matchers
 
 class JavaLibraryCrossProjectTargetJvmVersionIntegrationTest extends AbstractIntegrationSpec {
     ResolveTestFixture resolve
@@ -55,33 +56,26 @@ class JavaLibraryCrossProjectTargetJvmVersionIntegrationTest extends AbstractInt
         fails ':checkDeps'
 
         then:
-        failure.assertHasCause('''No matching variant of project :producer was found. The consumer was configured to find the public API of a library for use during compile-time, compatible with Java 6, preferably in the form of class files, preferably optimized for standard JVMs, and its dependencies declared externally but:
-  - Variant 'apiElements' declares the public API of a library for use during compile-time, packaged as a jar, and its dependencies declared externally:
+        failure.assertHasCause('''No matching variant of project :producer was found. The consumer was configured to find a library for use during compile-time, compatible with Java 6, preferably in the form of class files, preferably optimized for standard JVMs, and its dependencies declared externally but:
+  - Variant 'apiElements' declares a library for use during compile-time, packaged as a jar, and its dependencies declared externally:
       - Incompatible because this component declares a component, compatible with Java 7 and the consumer needed a component, compatible with Java 6
       - Other compatible attribute:
           - Doesn't say anything about its target Java environment (preferred optimized for standard JVMs)
   - Variant 'mainSourceElements' declares a component, and its dependencies declared externally:
       - Incompatible because this component declares a component of category 'verification' and the consumer needed a library
       - Other compatible attributes:
-          - Doesn't say anything about its API type (required the public API)
           - Doesn't say anything about its elements (required them preferably in the form of class files)
           - Doesn't say anything about its target Java environment (preferred optimized for standard JVMs)
           - Doesn't say anything about its target Java version (required compatibility with Java 6)
           - Doesn't say anything about its usage (required compile-time)
-  - Variant 'privateApiElements' declares the private API of a library for use during compile-time, packaged as a jar, and its dependencies declared externally:
-      - Incompatible because this component declares a component, compatible with Java 7 and the consumer needed a component, compatible with Java 6
-      - Other compatible attribute:
-          - Doesn't say anything about its target Java environment (preferred optimized for standard JVMs)
   - Variant 'runtimeElements' declares a library for use during runtime, packaged as a jar, and its dependencies declared externally:
       - Incompatible because this component declares a component, compatible with Java 7 and the consumer needed a component, compatible with Java 6
-      - Other compatible attributes:
-          - Doesn't say anything about its API type (required the public API)
+      - Other compatible attribute:
           - Doesn't say anything about its target Java environment (preferred optimized for standard JVMs)
   - Variant 'testResultsElementsForTest':
       - Incompatible because this component declares a component of category 'verification' and the consumer needed a library
       - Other compatible attributes:
           - Doesn't say anything about how its dependencies are found (required its dependencies declared externally)
-          - Doesn't say anything about its API type (required the public API)
           - Doesn't say anything about its elements (required them preferably in the form of class files)
           - Doesn't say anything about its target Java environment (preferred optimized for standard JVMs)
           - Doesn't say anything about its target Java version (required compatibility with Java 6)
@@ -93,7 +87,6 @@ class JavaLibraryCrossProjectTargetJvmVersionIntegrationTest extends AbstractInt
             // avoid test noise so that typically version 8 is not selected when running on JDK 8
             configurations.apiElements.attributes.attribute(TargetJvmVersion.TARGET_JVM_VERSION_ATTRIBUTE, 1000)
             configurations.runtimeElements.attributes.attribute(TargetJvmVersion.TARGET_JVM_VERSION_ATTRIBUTE, 1000)
-            configurations.privateApiElements.attributes.attribute(TargetJvmVersion.TARGET_JVM_VERSION_ATTRIBUTE, 1000)
 
             [6, 7, 9].each { v ->
                 configurations {
@@ -102,7 +95,6 @@ class JavaLibraryCrossProjectTargetJvmVersionIntegrationTest extends AbstractInt
                         canBeResolved = false
                         attributes {
                             attribute(Usage.USAGE_ATTRIBUTE, objects.named(Usage, 'java-api'))
-                            attribute(ApiType.TYPE_ATTRIBUTE, objects.named(ApiType, ApiType.PUBLIC))
                             attribute(LibraryElements.LIBRARY_ELEMENTS_ATTRIBUTE, objects.named(LibraryElements, 'jar'))
                             attribute(Bundling.BUNDLING_ATTRIBUTE, objects.named(Bundling, 'external'))
                             attribute(TargetJvmVersion.TARGET_JVM_VERSION_ATTRIBUTE, v)
@@ -129,7 +121,6 @@ class JavaLibraryCrossProjectTargetJvmVersionIntegrationTest extends AbstractInt
                             'org.gradle.dependency.bundling': 'external',
                             'org.gradle.jvm.version': selected,
                             'org.gradle.usage':'java-api',
-                            'org.gradle.api-type': 'public',
                             'org.gradle.libraryelements': 'jar'
                     ])
                     artifact(name: "producer-jdk${selected}")
@@ -164,37 +155,7 @@ class JavaLibraryCrossProjectTargetJvmVersionIntegrationTest extends AbstractInt
         fails ':checkDeps'
 
         then:
-        failure.assertHasCause("""No matching variant of project :producer was found. The consumer was configured to find the public API of a library for use during compile-time, compatible with Java 6, preferably in the form of class files, preferably optimized for standard JVMs, and its dependencies declared externally but:
-  - Variant 'apiElements' declares the public API of a library for use during compile-time, packaged as a jar, and its dependencies declared externally:
-      - Incompatible because this component declares a component, compatible with Java 7 and the consumer needed a component, compatible with Java 6
-      - Other compatible attribute:
-          - Doesn't say anything about its target Java environment (preferred optimized for standard JVMs)
-  - Variant 'mainSourceElements' declares a component, and its dependencies declared externally:
-      - Incompatible because this component declares a component of category 'verification' and the consumer needed a library
-      - Other compatible attributes:
-          - Doesn't say anything about its API type (required the public API)
-          - Doesn't say anything about its elements (required them preferably in the form of class files)
-          - Doesn't say anything about its target Java environment (preferred optimized for standard JVMs)
-          - Doesn't say anything about its target Java version (required compatibility with Java 6)
-          - Doesn't say anything about its usage (required compile-time)
-  - Variant 'privateApiElements' declares the private API of a library for use during compile-time, packaged as a jar, and its dependencies declared externally:
-      - Incompatible because this component declares a component, compatible with Java 7 and the consumer needed a component, compatible with Java 6
-      - Other compatible attribute:
-          - Doesn't say anything about its target Java environment (preferred optimized for standard JVMs)
-  - Variant 'runtimeElements' declares a library for use during runtime, packaged as a jar, and its dependencies declared externally:
-      - Incompatible because this component declares a component, compatible with Java 7 and the consumer needed a component, compatible with Java 6
-      - Other compatible attributes:
-          - Doesn't say anything about its API type (required the public API)
-          - Doesn't say anything about its target Java environment (preferred optimized for standard JVMs)
-  - Variant 'testResultsElementsForTest':
-      - Incompatible because this component declares a component of category 'verification' and the consumer needed a library
-      - Other compatible attributes:
-          - Doesn't say anything about how its dependencies are found (required its dependencies declared externally)
-          - Doesn't say anything about its API type (required the public API)
-          - Doesn't say anything about its elements (required them preferably in the form of class files)
-          - Doesn't say anything about its target Java environment (preferred optimized for standard JVMs)
-          - Doesn't say anything about its target Java version (required compatibility with Java 6)
-          - Doesn't say anything about its usage (required compile-time)""")
+        failure.assertThatCause(Matchers.startsWith("No matching variant of project :producer was found."))
 
         when:
         buildFile << """
@@ -213,8 +174,7 @@ class JavaLibraryCrossProjectTargetJvmVersionIntegrationTest extends AbstractInt
                             'org.gradle.dependency.bundling': 'external',
                             'org.gradle.jvm.version': 7,
                             'org.gradle.usage':'java-api',
-                            'org.gradle.libraryelements': 'jar',
-                            'org.gradle.api-type': 'public'
+                            'org.gradle.libraryelements': 'jar'
                     ])
                     artifact name: 'main', extension: '', type: 'java-classes-directory'
                 }
