+++ platforms/jvm/plugins-java-library/src/integTest/groovy/org/gradle/java/JavaLibraryOutgoingVariantsIntegrationTest.groovy
@@ -122,42 +122,12 @@ project(':consumer') {
         """)
     }
 
-    def "provides API compile jar variant - requestViewAttribute: #requestViewAttribute"() {
-        buildFile << """
-            project(':consumer') {
-                apply plugin: 'jvm-ecosystem'
-                configurations.consume.attributes.attribute(Usage.USAGE_ATTRIBUTE, objects.named(Usage, Usage.JAVA_API))
-                if ($requestViewAttribute) {
-                    configurations.consume.attributes.attribute(ApiType.TYPE_ATTRIBUTE, objects.named(ApiType, ApiType.PUBLIC))
-                }
-            }
-        """
-
-        when:
-        resolve()
-
-        then:
-        result.assertTasksExecuted(":other-java:compileJava", ":java:compileJava", ":java:processResources", ":java:classes", ":java:jar", ":consumer:resolve")
-        assertResolveOutput("""
-            files: [java.jar, api-1.0.jar, compile-only-api-1.0.jar]
-            java.jar (project :java) {artifactType=jar, org.gradle.api-type=public, org.gradle.category=library, org.gradle.dependency.bundling=external, ${defaultTargetPlatform()}, org.gradle.libraryelements=jar, org.gradle.usage=java-api}
-            api-1.0.jar (test:api:1.0) {artifactType=jar, org.gradle.category=library, org.gradle.libraryelements=jar, org.gradle.status=release, org.gradle.usage=java-api}
-            compile-only-api-1.0.jar (test:compile-only-api:1.0) {artifactType=jar, org.gradle.category=library, org.gradle.libraryelements=jar, org.gradle.status=release, org.gradle.usage=java-api}
-        """)
-
-        where:
-        requestViewAttribute << [true, false]
-    }
-
-    def "provides API compile classes variant - requestViewAttribute: #requestViewAttribute"() {
+    def "provides API classes variant"() {
         buildFile << """
             project(':consumer') {
                 apply plugin: 'jvm-ecosystem'
                 configurations.consume.attributes.attribute(Usage.USAGE_ATTRIBUTE, objects.named(Usage, Usage.JAVA_API))
                 configurations.consume.attributes.attribute(LibraryElements.LIBRARY_ELEMENTS_ATTRIBUTE, objects.named(LibraryElements, LibraryElements.CLASSES))
-                if ($requestViewAttribute) {
-                    configurations.consume.attributes.attribute(ApiType.TYPE_ATTRIBUTE, objects.named(ApiType, ApiType.PUBLIC))
-                }
             }
         """
 
@@ -168,77 +138,10 @@ project(':consumer') {
         result.assertTasksExecuted(":other-java:compileJava", ":java:compileJava", ":consumer:resolve")
         assertResolveOutput("""
             files: [main, api-1.0.jar, compile-only-api-1.0.jar]
-            main (project :java) {artifactType=java-classes-directory, org.gradle.api-type=public, org.gradle.category=library, org.gradle.dependency.bundling=external, ${defaultTargetPlatform()}, org.gradle.libraryelements=classes, org.gradle.usage=java-api}
+            main (project :java) {artifactType=java-classes-directory, org.gradle.category=library, org.gradle.dependency.bundling=external, ${defaultTargetPlatform()}, org.gradle.libraryelements=classes, org.gradle.usage=java-api}
             api-1.0.jar (test:api:1.0) {artifactType=jar, org.gradle.category=library, org.gradle.libraryelements=jar, org.gradle.status=release, org.gradle.usage=java-api}
             compile-only-api-1.0.jar (test:compile-only-api:1.0) {artifactType=jar, org.gradle.category=library, org.gradle.libraryelements=jar, org.gradle.status=release, org.gradle.usage=java-api}
         """)
-
-        where:
-        requestViewAttribute << [true, false]
-    }
-
-    def "provides implementation compile JAR variant - requestJarAttribute: #requestJarAttribute"() {
-        buildFile << """
-            project(':consumer') {
-                apply plugin: 'jvm-ecosystem'
-                configurations.consume.attributes.attribute(Usage.USAGE_ATTRIBUTE, objects.named(Usage, Usage.JAVA_API))
-                configurations.consume.attributes.attribute(ApiType.TYPE_ATTRIBUTE, objects.named(ApiType, ApiType.PRIVATE))
-                if ($requestJarAttribute) {
-                    configurations.consume.attributes.attribute(LibraryElements.LIBRARY_ELEMENTS_ATTRIBUTE, objects.named(LibraryElements, LibraryElements.JAR))
-                }
-            }
-        """
-
-        when:
-        resolve()
-
-        then:
-        result.assertTasksExecuted(":java:classes", ":java:compileJava", ":java:jar", ":java:processResources", ":other-java:classes", ":other-java:compileJava", ":other-java:jar", ":other-java:processResources", ":consumer:resolve")
-        assertResolveOutput("""
-            files: [java.jar, file-dep.jar, api-1.0.jar, compile-only-1.0.jar, compile-only-api-1.0.jar, other-java.jar, implementation-1.0.jar]
-            java.jar (project :java) {artifactType=jar, org.gradle.api-type=private, org.gradle.category=library, org.gradle.dependency.bundling=external, ${defaultTargetPlatform()}, org.gradle.libraryelements=jar, org.gradle.usage=java-api}
-            file-dep.jar {artifactType=jar, org.gradle.libraryelements=jar, org.gradle.usage=java-runtime}
-            api-1.0.jar (test:api:1.0) {artifactType=jar, org.gradle.category=library, org.gradle.libraryelements=jar, org.gradle.status=release, org.gradle.usage=java-api}
-            compile-only-1.0.jar (test:compile-only:1.0) {artifactType=jar, org.gradle.category=library, org.gradle.libraryelements=jar, org.gradle.status=release, org.gradle.usage=java-api}
-            compile-only-api-1.0.jar (test:compile-only-api:1.0) {artifactType=jar, org.gradle.category=library, org.gradle.libraryelements=jar, org.gradle.status=release, org.gradle.usage=java-api}
-            other-java.jar (project :other-java) {artifactType=jar, org.gradle.api-type=private, org.gradle.category=library, org.gradle.dependency.bundling=external, ${defaultTargetPlatform()}, org.gradle.libraryelements=jar, org.gradle.usage=java-api}
-            implementation-1.0.jar (test:implementation:1.0) {artifactType=jar, org.gradle.category=library, org.gradle.libraryelements=jar, org.gradle.status=release, org.gradle.usage=java-api}
-        """)
-
-        where:
-        requestJarAttribute << [true, false]
-    }
-
-    def "provides implementation compile classes and resources variant"() {
-        buildFile << """
-            project(':consumer') {
-                apply plugin: 'jvm-ecosystem'
-                configurations.consume.attributes.attribute(Usage.USAGE_ATTRIBUTE, objects.named(Usage, Usage.JAVA_API))
-                configurations.consume.attributes.attribute(ApiType.TYPE_ATTRIBUTE, objects.named(ApiType, ApiType.PRIVATE))
-                configurations.consume.attributes.attribute(LibraryElements.LIBRARY_ELEMENTS_ATTRIBUTE, objects.named(LibraryElements, LibraryElements.${requestedElements}))
-            }
-        """
-
-        when:
-        resolve()
-
-        then:
-        result.assertTasksExecuted(":consumer:resolve", ":java:classes", ":java:compileJava", ":java:processResources", ":other-java:classes", ":other-java:compileJava", ":other-java:processResources")
-        assertResolveOutput("""
-            files: [main, main, file-dep.jar, api-1.0.jar, compile-only-1.0.jar, compile-only-api-1.0.jar, main, main, implementation-1.0.jar]
-            main (project :java) {artifactType=directory, org.gradle.api-type=private, org.gradle.category=library, org.gradle.dependency.bundling=external, ${defaultTargetPlatform()}, org.gradle.libraryelements=classes+resources, org.gradle.usage=java-api}
-            main (project :java) {artifactType=directory, org.gradle.api-type=private, org.gradle.category=library, org.gradle.dependency.bundling=external, ${defaultTargetPlatform()}, org.gradle.libraryelements=classes+resources, org.gradle.usage=java-api}
-            file-dep.jar {artifactType=jar, org.gradle.libraryelements=jar, org.gradle.usage=java-runtime}
-            api-1.0.jar (test:api:1.0) {artifactType=jar, org.gradle.category=library, org.gradle.libraryelements=jar, org.gradle.status=release, org.gradle.usage=java-api}
-            compile-only-1.0.jar (test:compile-only:1.0) {artifactType=jar, org.gradle.category=library, org.gradle.libraryelements=jar, org.gradle.status=release, org.gradle.usage=java-api}
-            compile-only-api-1.0.jar (test:compile-only-api:1.0) {artifactType=jar, org.gradle.category=library, org.gradle.libraryelements=jar, org.gradle.status=release, org.gradle.usage=java-api}
-            main (project :other-java) {artifactType=directory, org.gradle.api-type=private, org.gradle.category=library, org.gradle.dependency.bundling=external, ${defaultTargetPlatform()}, org.gradle.libraryelements=classes+resources, org.gradle.usage=java-api}
-            main (project :other-java) {artifactType=directory, org.gradle.api-type=private, org.gradle.category=library, org.gradle.dependency.bundling=external, ${defaultTargetPlatform()}, org.gradle.libraryelements=classes+resources, org.gradle.usage=java-api}
-            implementation-1.0.jar (test:implementation:1.0) {artifactType=jar, org.gradle.category=library, org.gradle.libraryelements=jar, org.gradle.status=release, org.gradle.usage=java-api}
-        """)
-
-        where:
-        requestedElements << ["CLASSES", "RESOURCES", "CLASSES_AND_RESOURCES"]
     }
 
     def "provides runtime jar variant - requestJarAttribute: #requestJarAttribute"() {
