+++ platforms/jvm/plugins-java-library/src/integTest/groovy/org/gradle/java/JavaLibraryPublishedTargetJvmVersionIntegrationTest.groovy
@@ -98,36 +98,30 @@ class JavaLibraryPublishedTargetJvmVersionIntegrationTest extends AbstractHttpDe
         fails ':checkDeps'
 
         then:
-        failure.assertHasCause('''No matching variant of org:producer:1.0 was found. The consumer was configured to find the public API of a library for use during compile-time, compatible with Java 5, preferably in the form of class files, preferably optimized for standard JVMs, and its dependencies declared externally but:
+        failure.assertHasCause('''No matching variant of org:producer:1.0 was found. The consumer was configured to find a library for use during compile-time, compatible with Java 5, preferably in the form of class files, preferably optimized for standard JVMs, and its dependencies declared externally but:
   - Variant 'apiElementsJdk6' declares a library for use during compile-time, packaged as a jar, and its dependencies declared externally:
       - Incompatible because this component declares a component, compatible with Java 6 and the consumer needed a component, compatible with Java 5
-      - Other compatible attributes:
-          - Doesn't say anything about its API type (required the public API)
+      - Other compatible attribute:
           - Doesn't say anything about its target Java environment (preferred optimized for standard JVMs)
   - Variant 'apiElementsJdk7' declares a library for use during compile-time, packaged as a jar, and its dependencies declared externally:
       - Incompatible because this component declares a component, compatible with Java 7 and the consumer needed a component, compatible with Java 5
-      - Other compatible attributes:
-          - Doesn't say anything about its API type (required the public API)
+      - Other compatible attribute:
           - Doesn't say anything about its target Java environment (preferred optimized for standard JVMs)
   - Variant 'apiElementsJdk9' declares a library for use during compile-time, packaged as a jar, and its dependencies declared externally:
       - Incompatible because this component declares a component, compatible with Java 9 and the consumer needed a component, compatible with Java 5
-      - Other compatible attributes:
-          - Doesn't say anything about its API type (required the public API)
+      - Other compatible attribute:
           - Doesn't say anything about its target Java environment (preferred optimized for standard JVMs)
   - Variant 'runtimeElementsJdk6' declares a library for use during runtime, packaged as a jar, and its dependencies declared externally:
       - Incompatible because this component declares a component, compatible with Java 6 and the consumer needed a component, compatible with Java 5
-      - Other compatible attributes:
-          - Doesn't say anything about its API type (required the public API)
+      - Other compatible attribute:
           - Doesn't say anything about its target Java environment (preferred optimized for standard JVMs)
   - Variant 'runtimeElementsJdk7' declares a library for use during runtime, packaged as a jar, and its dependencies declared externally:
       - Incompatible because this component declares a component, compatible with Java 7 and the consumer needed a component, compatible with Java 5
-      - Other compatible attributes:
-          - Doesn't say anything about its API type (required the public API)
+      - Other compatible attribute:
           - Doesn't say anything about its target Java environment (preferred optimized for standard JVMs)
   - Variant 'runtimeElementsJdk9' declares a library for use during runtime, packaged as a jar, and its dependencies declared externally:
       - Incompatible because this component declares a component, compatible with Java 9 and the consumer needed a component, compatible with Java 5
-      - Other compatible attributes:
-          - Doesn't say anything about its API type (required the public API)
+      - Other compatible attribute:
           - Doesn't say anything about its target Java environment (preferred optimized for standard JVMs)''')
     }
