+++ platforms/jvm/plugins-java/src/main/java/org/gradle/api/plugins/JavaPlugin.java
@@ -48,12 +48,10 @@ import org.gradle.api.tasks.TaskContainer;
 import org.gradle.api.tasks.bundling.Jar;
 import org.gradle.api.tasks.diagnostics.DependencyInsightReportTask;
 import org.gradle.api.tasks.testing.Test;
-import org.gradle.internal.deprecation.DeprecationLogger;
 import org.gradle.internal.execution.BuildOutputCleanupRegistry;
 import org.gradle.jvm.component.internal.DefaultJvmSoftwareComponent;
 import org.gradle.jvm.component.internal.JvmSoftwareComponentInternal;
 import org.gradle.testing.base.TestingExtension;
-import org.gradle.util.GradleVersion;
 
 import javax.inject.Inject;
 import java.util.Collections;
@@ -277,7 +275,6 @@ public abstract class JavaPlugin implements Plugin<Project> {
         configureTestTaskOrdering(project.getTasks());
         configureDiagnostics(project, javaComponent.getMainFeature());
         configureBuild(project);
-        disablePrivateApiElementsForKmp(project, javaComponent.getMainFeature());
     }
 
     private static JvmFeatureInternal createMainFeature(ProjectInternal project, SourceSetContainer sourceSets) {
@@ -309,10 +306,6 @@ public abstract class JavaPlugin implements Plugin<Project> {
         // TODO: This process of manually adding variants to the component should be handled automatically when adding the feature to the component.
         component.addVariantsFromConfiguration(mainFeature.getApiElementsConfiguration(), new JavaConfigurationVariantMapping("compile", false, mainFeature.getCompileClasspathConfiguration()));
         component.addVariantsFromConfiguration(mainFeature.getRuntimeElementsConfiguration(), new JavaConfigurationVariantMapping("runtime", false, mainFeature.getRuntimeClasspathConfiguration()));
-
-        // Note: `privateApiElements` is intentionally not published!
-        // It is an internal detail of the component and should only be accessible within the component.
-        // Until we implement some variant visibility model, we should not publish this variant.
 
         // Create the default test suite
         JvmTestSuite defaultTestSuite = createDefaultTestSuite(mainFeature, project.getConfigurations(), project.getTasks(), project.getExtensions(), project.getObjects());
@@ -424,21 +417,5 @@ public abstract class JavaPlugin implements Plugin<Project> {
         Project project = task.getProject();
         final Configuration configuration = project.getConfigurations().getByName(configurationName);
         task.dependsOn(configuration.getTaskDependencyFromProjectDependency(useDependedOn, otherProjectTaskName));
-    }
-
-    /**
-     * Older versions of KMP, when using the {@code withJava} method of the jvm target, intentionally add
-     * configurations which are ambiguous with the ones added by this plugin. They get around the ambiguity
-     * by setting the configurations to not be consumable. These old versions are not aware of this configuration
-     * and thus are not able to set it to non-consumable.
-     *
-     * We maintain compatibility with these old kmp versions by setting this configuration to non-consumable ourselves.
-     */
-    private static void disablePrivateApiElementsForKmp(Project project, JvmFeatureInternal feature) {
-        assert GradleVersion.current().compareTo(GradleVersion.version("9.0")) < 0; // Delete this method in 9.0
-        project.getPlugins().withId("org.jetbrains.kotlin.multiplatform", a -> {
-            Configuration conf = feature.getPrivateApiElementsConfiguration();
-            DeprecationLogger.whileDisabled(() -> conf.setCanBeConsumed(false));
-        });
     }
 }
