+++ platforms/software/dependency-management/src/integTest/groovy/org/gradle/integtests/resolve/platforms/PlatformResolveIntegrationTest.groovy
@@ -184,8 +184,7 @@ class PlatformResolveIntegrationTest extends AbstractHttpDependencyResolutionTes
                                             'org.gradle.dependency.bundling': 'external',
                                             'org.gradle.jvm.version': JavaVersion.current().majorVersion,
                                             'org.gradle.usage': 'java-api',
-                                            'org.gradle.libraryelements': 'jar',
-                                            'org.gradle.api-type': 'public'])
+                                            'org.gradle.libraryelements': 'jar'])
                     constraint("org:platform:1.0", "org:platform:1.0") {
                         variant("platform-compile", [
                             'org.gradle.usage': 'java-api',
