+++ subprojects/diagnostics/src/integTest/groovy/org/gradle/api/tasks/diagnostics/DependencyInsightReportVariantDetailsIntegrationTest.groovy
@@ -59,7 +59,6 @@ class DependencyInsightReportVariantDetailsIntegrationTest extends AbstractInteg
         then:
         outputContains """project :b
 ${variantOf("apiElements", [
-            "org.gradle.api-type": of("public", "public"),
             "org.gradle.category": of("library", "library"),
             "org.gradle.dependency.bundling": of("external", "external"),
             "org.gradle.jvm.version": of(JavaVersion.current().majorVersion, JavaVersion.current().majorVersion),
@@ -141,7 +140,6 @@ Selected Variant(s)
 -------------------
 
 ${variantOf('apiElements', [
-                "org.gradle.api-type": of('public', 'public'),
                 'org.gradle.category': of('library', 'library'),
                 'org.gradle.dependency.bundling': of('external', 'external'),
                 'org.gradle.jvm.version': of(JavaVersion.current().majorVersion, JavaVersion.current().majorVersion),
@@ -155,7 +153,6 @@ Unselected Variant(s)
 ---------------------
 
 ${variantOf('apiElements-classes', [
-                'org.gradle.api-type': of('public', 'public'),
                 'org.gradle.category': of('library', 'library'),
                 'org.gradle.dependency.bundling': of('external', 'external'),
                 'org.gradle.jvm.version': of(JavaVersion.current().majorVersion, JavaVersion.current().majorVersion),
@@ -168,32 +165,10 @@ ${variantOf('mainSourceElements', [
                 'org.gradle.verificationtype': of('main-sources', ''),
                 'org.gradle.category': of('verification', 'library'),
                 'org.gradle.dependency.bundling': of('external', 'external'),
-                'org.gradle.api-type': of('', 'public'),
                 'org.gradle.jvm.environment': of('', 'standard-jvm'),
                 'org.gradle.jvm.version': of('', JavaVersion.current().majorVersion),
                 'org.gradle.libraryelements': of('', 'classes'),
                 'org.gradle.usage': of('', 'java-api'),
-            ])}
-
-${variantOf('privateApiElements', [
-                artifactType: of('jar', ''),
-                'org.gradle.api-type': of('private', 'public'),
-                'org.gradle.category': of('library', 'library'),
-                'org.gradle.dependency.bundling': of('external', 'external'),
-                'org.gradle.jvm.version': of(JavaVersion.current().majorVersion, JavaVersion.current().majorVersion),
-                'org.gradle.libraryelements': of('jar', 'classes'),
-                'org.gradle.usage': of('java-api', 'java-api'),
-                'org.gradle.jvm.environment': of('', 'standard-jvm')
-            ])}
-
-${variantOf('privateApiElements-classesAndResources', [
-                'org.gradle.api-type': of('private', 'public'),
-                'org.gradle.category': of('library', 'library'),
-                'org.gradle.dependency.bundling': of('external', 'external'),
-                'org.gradle.jvm.version': of(JavaVersion.current().majorVersion, JavaVersion.current().majorVersion),
-                'org.gradle.libraryelements': of('classes+resources', 'classes'),
-                'org.gradle.usage': of('java-api', 'java-api'),
-                'org.gradle.jvm.environment': of('', 'standard-jvm')
             ])}
 
 ${variantOf('runtimeElements', [
@@ -203,7 +178,6 @@ ${variantOf('runtimeElements', [
                 'org.gradle.jvm.version': of(JavaVersion.current().majorVersion, JavaVersion.current().majorVersion),
                 'org.gradle.libraryelements': of('jar', 'classes'),
                 'org.gradle.usage': of('java-runtime', 'java-api'),
-                'org.gradle.api-type': of('', 'public'),
                 'org.gradle.jvm.environment': of('', 'standard-jvm'),
             ])}
 
@@ -213,7 +187,6 @@ ${variantOf('runtimeElements-classes', [
                 'org.gradle.jvm.version': of(JavaVersion.current().majorVersion, JavaVersion.current().majorVersion),
                 'org.gradle.libraryelements': of('classes', 'classes'),
                 'org.gradle.usage': of('java-runtime', 'java-api'),
-                'org.gradle.api-type': of('', 'public'),
                 'org.gradle.jvm.environment': of('', 'standard-jvm'),
             ])}
 
@@ -223,7 +196,6 @@ ${variantOf('runtimeElements-resources', [
                 'org.gradle.jvm.version': of(JavaVersion.current().majorVersion, JavaVersion.current().majorVersion),
                 'org.gradle.libraryelements': of('resources', 'classes'),
                 'org.gradle.usage': of('java-runtime', 'java-api'),
-                'org.gradle.api-type': of('', 'public'),
                 'org.gradle.jvm.environment': of('', 'standard-jvm'),
             ])}
 
@@ -233,7 +205,6 @@ ${variantOf('testResultsElementsForTest', [
                 'org.gradle.testsuite.type': of('unit-test', ''),
                 'org.gradle.verificationtype': of('test-results', ''),
                 'org.gradle.category': of('verification', 'library'),
-                'org.gradle.api-type': of('', 'public'),
                 'org.gradle.dependency.bundling': of('', 'external'),
                 'org.gradle.jvm.environment': of('', 'standard-jvm'),
                 'org.gradle.jvm.version': of('', JavaVersion.current().majorVersion),
@@ -296,7 +267,6 @@ Unselected Variant(s)
 
 ${variantOf('apiElements', [
                 artifactType: of('jar', ''),
-                'org.gradle.api-type': of('public', ''),
                 'org.gradle.category': of('library', 'library'),
                 'org.gradle.dependency.bundling': of('external', 'external'),
                 'org.gradle.jvm.version': of(JavaVersion.current().majorVersion, JavaVersion.current().majorVersion),
@@ -306,7 +276,6 @@ ${variantOf('apiElements', [
             ])}
 
 ${variantOf("apiElements-classes", [
-                'org.gradle.api-type': of('public', ''),
                 'org.gradle.category': of('library', 'library'),
                 'org.gradle.dependency.bundling': of('external', 'external'),
                 'org.gradle.jvm.version': of(JavaVersion.current().majorVersion, JavaVersion.current().majorVersion),
@@ -323,27 +292,6 @@ ${variantOf('mainSourceElements', [
                 'org.gradle.jvm.version': of('', JavaVersion.current().majorVersion),
                 'org.gradle.libraryelements': of('', 'jar'),
                 'org.gradle.usage': of('', 'java-runtime'),
-            ])}
-
-${variantOf('privateApiElements', [
-                artifactType: of('jar', ''),
-                'org.gradle.api-type': of('private', ''),
-                'org.gradle.category': of('library', 'library'),
-                'org.gradle.dependency.bundling': of('external', 'external'),
-                'org.gradle.jvm.version': of(JavaVersion.current().majorVersion, JavaVersion.current().majorVersion),
-                'org.gradle.libraryelements': of('jar', 'jar'),
-                'org.gradle.usage': of('java-api', 'java-runtime'),
-                'org.gradle.jvm.environment': of('', 'standard-jvm')
-            ])}
-
-${variantOf('privateApiElements-classesAndResources', [
-                'org.gradle.api-type': of('private', ''),
-                'org.gradle.category': of('library', 'library'),
-                'org.gradle.dependency.bundling': of('external', 'external'),
-                'org.gradle.jvm.version': of(JavaVersion.current().majorVersion, JavaVersion.current().majorVersion),
-                'org.gradle.libraryelements': of('classes+resources', 'jar'),
-                'org.gradle.usage': of('java-api', 'java-runtime'),
-                'org.gradle.jvm.environment': of('', 'standard-jvm')
             ])}
 
 ${variantOf('runtimeElements-classes', [
@@ -449,7 +397,6 @@ project :$expectedProject
     | org.gradle.category            | library             | library      |
     | org.gradle.libraryelements     | jar                 | classes      |
     | org.gradle.usage               | java-api            | java-api     |
-    | org.gradle.api-type            |                     | public       |
     | org.gradle.blah                |                     | something    |
     | org.gradle.dependency.bundling |                     | external     |
     | org.gradle.jvm.environment     |                     | standard-jvm |
