+++ subprojects/diagnostics/src/integTest/groovy/org/gradle/api/tasks/diagnostics/OutgoingVariantsReportTaskIntegrationTest.groovy
@@ -229,14 +229,13 @@ Attributes
         def sourceMainResourcePath = file('src/main/resources').getRelativePathFromBase()
         def resultsBinPath = file('build/test-results/test/binary').getRelativePathFromBase()
         result.groupedOutput.task(":outgoingVariants").assertOutputContains """--------------------------------------------------
-Variant apiElements (i)
+Variant apiElements
 --------------------------------------------------
-Public API elements for the 'main' feature.
+API elements for the 'main' feature.
 
 Capabilities
     - org:myLib:1.0 (default capability)
 Attributes
-    - org.gradle.api-type            = public
     - org.gradle.category            = library
     - org.gradle.dependency.bundling = external
     - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
@@ -248,12 +247,11 @@ Artifacts
 Secondary Variants (*)
 
     --------------------------------------------------
-    Secondary Variant classes (i)
+    Secondary Variant classes
     --------------------------------------------------
     Directories containing compiled class files for main.
 
     Attributes
-        - org.gradle.api-type            = public
         - org.gradle.category            = library
         - org.gradle.dependency.bundling = external
         - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
@@ -296,41 +294,6 @@ Attributes
 Artifacts
     - $sourceMainJavaPath (artifactType = directory)
     - $sourceMainResourcePath (artifactType = directory)
-
---------------------------------------------------
-Variant privateApiElements (i)
---------------------------------------------------
-Private API elements for the 'main' feature.
-
-Capabilities
-    - org:myLib:1.0 (default capability)
-Attributes
-    - org.gradle.api-type            = private
-    - org.gradle.category            = library
-    - org.gradle.dependency.bundling = external
-    - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
-    - org.gradle.libraryelements     = jar
-    - org.gradle.usage               = java-api
-Artifacts
-    - $jarPath (artifactType = jar)
-
-Secondary Variants (*)
-
-    --------------------------------------------------
-    Secondary Variant classesAndResources (i)
-    --------------------------------------------------
-    Directories containing compiled class files and resources for main.
-
-    Attributes
-        - org.gradle.api-type            = private
-        - org.gradle.category            = library
-        - org.gradle.dependency.bundling = external
-        - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
-        - org.gradle.libraryelements     = classes+resources
-        - org.gradle.usage               = java-api
-    Artifacts
-        - $builtMainClassesPath (artifactType = directory)
-        - $builtMainResourcesPath (artifactType = directory)
 
 --------------------------------------------------
 Variant runtimeElements
@@ -424,14 +387,13 @@ Artifacts
         def sourceMainResourcePath = file('src/main/resources').getRelativePathFromBase()
         def resultsBinPath = file('build/test-results/test/binary').getRelativePathFromBase()
         result.groupedOutput.task(":outgoingVariants").assertOutputContains("""--------------------------------------------------
-Variant apiElements (i)
+Variant apiElements
 --------------------------------------------------
-Public API elements for the 'main' feature.
+API elements for the 'main' feature.
 
 Capabilities
     - org:myLib:1.0 (default capability)
 Attributes
-    - org.gradle.api-type            = public
     - org.gradle.category            = library
     - org.gradle.dependency.bundling = external
     - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
@@ -443,12 +405,11 @@ Artifacts
 Secondary Variants (*)
 
     --------------------------------------------------
-    Secondary Variant classes (i)
+    Secondary Variant classes
     --------------------------------------------------
     Directories containing compiled class files for main.
 
     Attributes
-        - org.gradle.api-type            = public
         - org.gradle.category            = library
         - org.gradle.dependency.bundling = external
         - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
@@ -506,41 +467,6 @@ Attributes
 Artifacts
     - $sourceMainJavaPath (artifactType = directory)
     - $sourceMainResourcePath (artifactType = directory)
-
---------------------------------------------------
-Variant privateApiElements (i)
---------------------------------------------------
-Private API elements for the 'main' feature.
-
-Capabilities
-    - org:myLib:1.0 (default capability)
-Attributes
-    - org.gradle.api-type            = private
-    - org.gradle.category            = library
-    - org.gradle.dependency.bundling = external
-    - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
-    - org.gradle.libraryelements     = jar
-    - org.gradle.usage               = java-api
-Artifacts
-    - $jarPath (artifactType = jar)
-
-Secondary Variants (*)
-
-    --------------------------------------------------
-    Secondary Variant classesAndResources (i)
-    --------------------------------------------------
-    Directories containing compiled class files and resources for main.
-
-    Attributes
-        - org.gradle.api-type            = private
-        - org.gradle.category            = library
-        - org.gradle.dependency.bundling = external
-        - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
-        - org.gradle.libraryelements     = classes+resources
-        - org.gradle.usage               = java-api
-    Artifacts
-        - $builtMainClassesPath (artifactType = directory)
-        - $builtMainResourcesPath (artifactType = directory)
 
 --------------------------------------------------
 Variant runtimeElements
@@ -648,14 +574,13 @@ Artifacts
         def sourceMainResourcePath = file('src/main/resources').getRelativePathFromBase()
         def resultsBinPath = file('build/test-results/test/binary').getRelativePathFromBase()
         result.groupedOutput.task(":outgoingVariants").assertOutputContains("""--------------------------------------------------
-Variant apiElements (i)
+Variant apiElements
 --------------------------------------------------
-Public API elements for the 'main' feature.
+API elements for the 'main' feature.
 
 Capabilities
     - org:myLib:1.0 (default capability)
 Attributes
-    - org.gradle.api-type            = public
     - org.gradle.category            = library
     - org.gradle.dependency.bundling = external
     - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
@@ -667,12 +592,11 @@ Artifacts
 Secondary Variants (*)
 
     --------------------------------------------------
-    Secondary Variant classes (i)
+    Secondary Variant classes
     --------------------------------------------------
     Directories containing compiled class files for main.
 
     Attributes
-        - org.gradle.api-type            = public
         - org.gradle.category            = library
         - org.gradle.dependency.bundling = external
         - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
@@ -730,41 +654,6 @@ Attributes
 Artifacts
     - $sourceMainJavaPath (artifactType = directory)
     - $sourceMainResourcePath (artifactType = directory)
-
---------------------------------------------------
-Variant privateApiElements (i)
---------------------------------------------------
-Private API elements for the 'main' feature.
-
-Capabilities
-    - org:myLib:1.0 (default capability)
-Attributes
-    - org.gradle.api-type            = private
-    - org.gradle.category            = library
-    - org.gradle.dependency.bundling = external
-    - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
-    - org.gradle.libraryelements     = jar
-    - org.gradle.usage               = java-api
-Artifacts
-    - $jarPath (artifactType = jar)
-
-Secondary Variants (*)
-
-    --------------------------------------------------
-    Secondary Variant classesAndResources (i)
-    --------------------------------------------------
-    Directories containing compiled class files and resources for main.
-
-    Attributes
-        - org.gradle.api-type            = private
-        - org.gradle.category            = library
-        - org.gradle.dependency.bundling = external
-        - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
-        - org.gradle.libraryelements     = classes+resources
-        - org.gradle.usage               = java-api
-    Artifacts
-        - $builtMainClassesPath (artifactType = directory)
-        - $builtMainResourcesPath (artifactType = directory)
 
 --------------------------------------------------
 Variant runtimeElements
@@ -935,14 +824,13 @@ Secondary Variants (*)
         def resultsBinPath = file('build/test-results/test/binary').getRelativePathFromBase()
 
         result.groupedOutput.task(":outgoingVariants").assertOutputContains("""--------------------------------------------------
-Variant apiElements (i)
+Variant apiElements
 --------------------------------------------------
-Public API elements for the 'main' feature.
+API elements for the 'main' feature.
 
 Capabilities
     - org:myLib:1.0 (default capability)
 Attributes
-    - org.gradle.api-type            = public
     - org.gradle.category            = library
     - org.gradle.dependency.bundling = external
     - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
@@ -954,12 +842,11 @@ Artifacts
 Secondary Variants (*)
 
     --------------------------------------------------
-    Secondary Variant classes (i)
+    Secondary Variant classes
     --------------------------------------------------
     Directories containing compiled class files for main.
 
     Attributes
-        - org.gradle.api-type            = public
         - org.gradle.category            = library
         - org.gradle.dependency.bundling = external
         - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
@@ -1002,41 +889,6 @@ Attributes
 Artifacts
     - $sourceMainJavaPath (artifactType = directory)
     - $sourceMainResourcePath (artifactType = directory)
-
---------------------------------------------------
-Variant privateApiElements (i)
---------------------------------------------------
-Private API elements for the 'main' feature.
-
-Capabilities
-    - org:myLib:1.0 (default capability)
-Attributes
-    - org.gradle.api-type            = private
-    - org.gradle.category            = library
-    - org.gradle.dependency.bundling = external
-    - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
-    - org.gradle.libraryelements     = jar
-    - org.gradle.usage               = java-api
-Artifacts
-    - $jarPath (artifactType = jar)
-
-Secondary Variants (*)
-
-    --------------------------------------------------
-    Secondary Variant classesAndResources (i)
-    --------------------------------------------------
-    Directories containing compiled class files and resources for main.
-
-    Attributes
-        - org.gradle.api-type            = private
-        - org.gradle.category            = library
-        - org.gradle.dependency.bundling = external
-        - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
-        - org.gradle.libraryelements     = classes+resources
-        - org.gradle.usage               = java-api
-    Artifacts
-        - $builtMainClassesPath (artifactType = directory)
-        - $builtMainResourcesPath (artifactType = directory)
 
 --------------------------------------------------
 Variant runtimeElements
@@ -1124,14 +976,13 @@ Artifacts
         def sourceMainResourcePath = file('src/main/resources').getRelativePathFromBase()
         def resultsBinPath = file('build/test-results/test/binary').getRelativePathFromBase()
         result.groupedOutput.task(":outgoingVariants").assertOutputContains("""--------------------------------------------------
-Variant apiElements (i)
+Variant apiElements
 --------------------------------------------------
-Public API elements for the 'main' feature.
+API elements for the 'main' feature.
 
 Capabilities
     - org:myLib:1.0 (default capability)
 Attributes
-    - org.gradle.api-type            = public
     - org.gradle.category            = library
     - org.gradle.dependency.bundling = external
     - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
@@ -1143,12 +994,11 @@ Artifacts
 Secondary Variants (*)
 
     --------------------------------------------------
-    Secondary Variant classes (i)
+    Secondary Variant classes
     --------------------------------------------------
     Directories containing compiled class files for main.
 
     Attributes
-        - org.gradle.api-type            = public
         - org.gradle.category            = library
         - org.gradle.dependency.bundling = external
         - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
@@ -1191,41 +1041,6 @@ Attributes
 Artifacts
     - $sourceMainJavaPath (artifactType = directory)
     - $sourceMainResourcePath (artifactType = directory)
-
---------------------------------------------------
-Variant privateApiElements (i)
---------------------------------------------------
-Private API elements for the 'main' feature.
-
-Capabilities
-    - org:myLib:1.0 (default capability)
-Attributes
-    - org.gradle.api-type            = private
-    - org.gradle.category            = library
-    - org.gradle.dependency.bundling = external
-    - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
-    - org.gradle.libraryelements     = jar
-    - org.gradle.usage               = java-api
-Artifacts
-    - $jarPath (artifactType = jar)
-
-Secondary Variants (*)
-
-    --------------------------------------------------
-    Secondary Variant classesAndResources (i)
-    --------------------------------------------------
-    Directories containing compiled class files and resources for main.
-
-    Attributes
-        - org.gradle.api-type            = private
-        - org.gradle.category            = library
-        - org.gradle.dependency.bundling = external
-        - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
-        - org.gradle.libraryelements     = classes+resources
-        - org.gradle.usage               = java-api
-    Artifacts
-        - $builtMainClassesPath (artifactType = directory)
-        - $builtMainResourcesPath (artifactType = directory)
 
 --------------------------------------------------
 Variant runtimeElements
@@ -1423,14 +1238,13 @@ Secondary Variants (*)
         def sourceMainResourcePath = file('src/main/resources').getRelativePathFromBase()
         def resultsBinPath = file('build/test-results/test/binary').getRelativePathFromBase()
         result.groupedOutput.task(":outgoingVariants").assertOutputContains("""--------------------------------------------------
-Variant apiElements (i)
+Variant apiElements
 --------------------------------------------------
-Public API elements for the 'main' feature.
+API elements for the 'main' feature.
 
 Capabilities
     - org:myLib:1.0 (default capability)
 Attributes
-    - org.gradle.api-type            = public
     - org.gradle.category            = library
     - org.gradle.dependency.bundling = external
     - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
@@ -1442,12 +1256,11 @@ Artifacts
 Secondary Variants (*)
 
     --------------------------------------------------
-    Secondary Variant classes (i)
+    Secondary Variant classes
     --------------------------------------------------
     Directories containing compiled class files for main.
 
     Attributes
-        - org.gradle.api-type            = public
         - org.gradle.category            = library
         - org.gradle.dependency.bundling = external
         - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
@@ -1490,41 +1303,6 @@ Attributes
 Artifacts
     - $sourceMainJavaPath (artifactType = directory)
     - $sourceMainResourcePath (artifactType = directory)
-
---------------------------------------------------
-Variant privateApiElements (i)
---------------------------------------------------
-Private API elements for the 'main' feature.
-
-Capabilities
-    - org:myLib:1.0 (default capability)
-Attributes
-    - org.gradle.api-type            = private
-    - org.gradle.category            = library
-    - org.gradle.dependency.bundling = external
-    - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
-    - org.gradle.libraryelements     = jar
-    - org.gradle.usage               = java-api
-Artifacts
-    - $jarPath (artifactType = jar)
-
-Secondary Variants (*)
-
-    --------------------------------------------------
-    Secondary Variant classesAndResources (i)
-    --------------------------------------------------
-    Directories containing compiled class files and resources for main.
-
-    Attributes
-        - org.gradle.api-type            = private
-        - org.gradle.category            = library
-        - org.gradle.dependency.bundling = external
-        - org.gradle.jvm.version         = ${JavaVersion.current().majorVersion}
-        - org.gradle.libraryelements     = classes+resources
-        - org.gradle.usage               = java-api
-    Artifacts
-        - $builtMainClassesPath (artifactType = directory)
-        - $builtMainResourcesPath (artifactType = directory)
 
 --------------------------------------------------
 Variant runtimeElements
