+++ subprojects/internal-integ-testing/src/main/groovy/org/gradle/test/fixtures/maven/MavenJavaModule.groovy
@@ -106,7 +106,6 @@ class MavenJavaModule extends DelegatingMavenModule<MavenFileModule> implements 
                                        "org.gradle.jvm.version": currentJavaVersion,
                                        "org.gradle.libraryelements": "jar",
                                        "org.gradle.usage": "java-api",
-                                       "org.gradle.api-type": "public"
                                       ])
         assertAttributes(runtimeElements, ["org.gradle.category": "library",
                                            "org.gradle.dependency.bundling": "external",
